refactor: extract LLM route core into callable generateLlmText
Move the three provider functions (Google/OpenAI/Anthropic) and their model
maps out of the route handler into a reusable core module. The POST handler
now delegates to the extracted functions with identical behaviour (existing
route tests unchanged), and the server-side workflow runner can call
generateLlmText directly instead of self-issuing an HTTP request.
